Fixed pools and protocol stat issues

+++ fathomapp-subgraph/src/position-manager.ts
@@ -29,7 +29,7 @@ export function newPositionHandler(event: LogNewPosition): void {
     position.collatralPoolName = poolId.toString()
     position.lockedCollateral = BigInt.fromString('0')
     position.debtShare = BigInt.fromString('0')
-    position.safetyBuffer= BigInt.fromString('1')
+    position.safetyBuffer= BigDecimal.fromString('1')
     position.safetyBufferInPrecent= BigDecimal.fromString('0')
     position.tvl = BigDecimal.fromString('0')
     position.positionStatus = 'active'
